Refactor per merge
