+++ src/extensions/default/analyzers/Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers.Common/SymbolExtensions.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Linq;
+using System.Threading;
 using Microsoft.CodeAnalysis;
 
 namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default
@@ -23,6 +25,65 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default
             {
                 return string.Equals(symbol.Name, name, StringComparison.OrdinalIgnoreCase);
             }
+        }
+
+        /// <summary>
+        /// Gets an existing parameter of the type from node if it is a method.
+        /// </summary>
+        /// <param name="node">A node that is potentially a method.</param>
+        /// <param name="semanticModel">The current semantic model.</param>
+        /// <param name="type">The type that should be matched.</param>
+        /// <param name="token">A cancellation token.</param>
+        /// <returns>A parameter symbol if a match is found.</returns>
+        public static IParameterSymbol? GetExistingParameterSymbol(this SyntaxNode node, SemanticModel semanticModel, ITypeSymbol? type, CancellationToken token)
+        {
+            if (type is null)
+            {
+                return null;
+            }
+
+            var symbol = semanticModel.GetDeclaredSymbol(node, token);
+
+            if (symbol is not IMethodSymbol method)
+            {
+                return null;
+            }
+
+            return method.Parameters.FirstOrDefault(p =>
+            {
+                if (p.Type is null)
+                {
+                    return false;
+                }
+
+                return SymbolEqualityComparer.IncludeNullability.Equals(p.Type, type);
+            });
+        }
+
+        /// <summary>
+        /// Gets a parent operation that satisfies the given predicate.
+        /// </summary>
+        /// <param name="operation">A starting operation.</param>
+        /// <param name="func">A predicate to check matches against.</param>
+        /// <returns>A matching operation if found.</returns>
+        public static IOperation? GetParentOperation(this IOperation? operation, Func<IOperation, bool> func)
+        {
+            if (func is null)
+            {
+                throw new ArgumentNullException(nameof(func));
+            }
+
+            while (operation is not null)
+            {
+                if (func(operation))
+                {
+                    return operation;
+                }
+
+                operation = operation.Parent;
+            }
+
+            return default;
         }
     }
 }
