+++ src/extensions/default/analyzers/Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers.Common/WorkspaceExtensions.cs
@@ -1,0 +1,34 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.CodeAnalysis;
+
+namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default
+{
+    public static class WorkspaceExtensions
+    {
+        public static bool TryGetDocument(this Solution sln, SyntaxTree? tree, [MaybeNullWhen(false)] out Document document)
+        {
+            if (tree is null || sln is null)
+            {
+                document = null;
+                return false;
+            }
+
+            foreach (var project in sln.Projects)
+            {
+                var doc = project.GetDocument(tree);
+
+                if (doc is not null)
+                {
+                    document = doc;
+                    return true;
+                }
+            }
+
+            document = null;
+            return false;
+        }
+    }
+}
